its working

+++ projects/rtos/task_management.h
@@ -24,7 +24,7 @@ struct task_block {
   int state;
   void (*start)(void * arg);
   void * arg;
-  uint32_t * sp;
+  uint8_t * sp;
 };
